big overhaul clean up of node property bin widgets.
