+++ NodeGraphQt/__init__.py
@@ -61,7 +61,7 @@ from .nodes.group_node import GroupNode
 from .widgets.node_widgets import NodeBaseWidget
 from .custom_widgets.nodes_tree import NodesTreeWidget
 from .custom_widgets.nodes_palette import NodesPaletteWidget
-from .custom_widgets.properties_bin import PropertiesBinWidget
+from .custom_widgets.properties_bin.node_property_widgets import PropertiesBinWidget
 
 
 __version__ = VERSION
