+++ NodeGraphQt/qgraphics/node_base.py
@@ -1024,4 +1024,4 @@ class NodeItem(AbstractNodeItem):
         widgets = node_dict.pop('widgets', {})
         for name, value in widgets.items():
             if self._widgets.get(name):
-                self._widgets[name].value = value
+                self._widgets[name].get_value = value
